### 2024-10-10 #262 @editorialbot
[PRE REVIEW]: A course on the setup, running, and analysis of biomolecular simulations
**Submitting author:** @degiacom (<a href="http://orcid.org/0000-0003-4672-471X">Matteo Degiacomi</a>)
**Repository:** https://github.com/CCPBioSim/BioSim-analysis-workshop
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
