### 2024-11-05 #265 @djcole56
Hi, I'll drop any issues with the notebooks in here as I'm going along.

In 4_Simulation_Setup, only this method works:
```
lig27_smiles = "[H:7][C@:6]1([C:13](=[C:11]([C:9](=[O:10])[O:8]1)[O:12][H:19])[O:14][H:20])[C@:3]([H:4])([C:2]([H:16])([H:17])[O:1][H:15])[O:5][H:18]"
ligand = Molecule.from_mapped_smiles(lig27_smiles)
```
ie reading from file does not (I guess because it's not a mapped smiles).

More generally, it would be good to include instruction for how to make a mapped smiles if this is necessary input.
