+++ README.md
@@ -69,7 +69,7 @@ A full list of the required Python packages can be seen inside [`environment.yml
 
 ## Authors
 
-- [Dr Matteo Degiacomi, Durham University](https://degiacomi.org)
+- [Dr Matteo Degiacomi, University of Edinburgh](https://degiacomi.org)
 - [Dr Antonia Mey, University of Edinburgh](https://mey-research.org/)
 - [Dr Micaela Matta, King's College London](https://www.kcl.ac.uk/people/micaela-matta)
 - Dr Richard Gowers
