+++ README.md
@@ -2,7 +2,7 @@
 
 ## Schedule
 
-The course is constituted of lectures (L1-8) and practical (P) sessions, subdivided in two units. Unit 1 is dedicated to providing foundations on protein structure and their preparation for MD simulation. Unit 2 is dedicated to describing means of extracting information from the molecular dynamics simulation of a protein.
+The course is constituted of lectures (L1-8) and practical (P) sessions, subdivided in two units. Unit 1 is dedicated to providing foundations on protein structure and their preparation for molecular dynamics (MD) simulation. Unit 2 is dedicated to describing means of extracting information from the MD simulation of a protein.
 
 ### Unit 1: Simulation Preparation
 
