### 2024-11-13 #265 @djcole56
@ppxasjsm @degiacom Finished going through the material. The course comprises a series of around 8 lectures and 8 workshops, with the first half devoted to setting up and running biomolecular simulations, and the second half to data analysis, including introductions to machine learning. The lectures are very clear and could be adopted by other teachers for their own courses. The workshops are presented through Jupyter notebooks, and I followed them on Google colab (as I imagine most students would). These are again very clearly presented and at a suitable level for a graduate or advanced undergrad class. Worked examples and problems are used throughout to hold attention, and the material usually progresses nicely from toy models to realistic simulation data. This course is commended for focusing on the fundamentals of simulation and analysis that is mostly agnostic to the underlying MD codes, and doesn't require any expensive licenses.

The material has clearly been used for teaching on several occasions, so I only found the one issue (above). Also nglview is not always able to run the trajectories on colab, but I this is a recurring problem, and not the authors' issue (it might still be worth warning students that they can use their favourite pdb viewer if this happens).

If I understand the journal requirements correctly, the GitHub page needs a 'statement of need' and 'community guidelines', then I'm happy to sign it off.
