### 2024-11-24 CCPBioSim/BioSim-analysis-workshop#4 @raquellrios
Repo intro
- Adding  the **target student** audience  (master, undergrad...) to the course description will increase it usability. I believe it will make it easier for students to determine if they have the right knowledge for the course, as well as for teachers to quickly assess if they can use the material.
- **Typo** in name of L5 and P5 title: Analyses --> Analysis
