### 2024-11-24 CCPBioSim/BioSim-analysis-workshop#5 @raquellrios
Comments L1
Comments:

- When talking about protein structure, it would be nice to just mention intrinsically disorder proteins. So that it is clear for students that proteins do not always adopt well folded structures overall. Personally, I found this later in my Biophysics course than I would have wanted, since disorder proteins are so interesting! I would maybe just add a sentence on **slide 20**.
- (Optional) On **slide 14** add  DSSP (Dictionary of Secondary Structure in Proteins) so that there is a bit more context for that abbreviation
- Maybe on **slide 31** where it says: "most contacts are short-lived" put an estimate of the time unit (e.g. in the order of nanoseconds) so that students get their head around how short that really is!
- I found the phrasing on **slide 34** a bit confusing, I think instead of "Proteins and diseases" I would say: "Proteins in diseases" (and similarly, 'Proteins in disorders' instead of 'Proteins and disorders).  I believe this clarifies the role of proteins in such cases and and make the slide easier to follow if not accompanied by a lecture.
- On **slide 44**  I would put the link to the PDB (currently only the citation is there), so that students can easily click it and inspect it by themselves.
- Love the slides with the brain and alphafold structure prediction ❤️
