### 2024-11-24 CCPBioSim/BioSim-analysis-workshop#6 @raquellrios
Comments L2
Comments:

- The QR code on **slide 4** is not working for me, but I assume it is deactivated for now.
- **Slide 17** has a great easy-to-follow explanation of what are pKa values. However I think it lacks a bullet point explaining what is the impact on simulations when protonation states are not properly predicted, such as, wrong binding affinity predictions.  Additionally, I would include a sentence saying: "finding the right protonation state for proteins and ligands can be challenging! "  Just so that students are aware of the importance and impact of protonation states.
- Typo on **slide 24**: "You have been told you to simulate" --> " You have been told to simulate"
- On the side of **slide 29**  I would add examples of non-protein molecules that are commonly found in the PDB (e.g., cholesterol, glycerol...)
