### 2024-11-24 CCPBioSim/BioSim-analysis-workshop#7 @raquellrios
Comments L3
Comments:

-  **Slide 7** should mention that the bound and docked ligands need to be sufficiently similar for shape-based docking methods to work effectively.
- (Optional) The image quality on **slide 9** is not very good, would it be possible to improve it?
- On **slide 15** I would clarify that the RMSD is measured relative to the crystal/reference structure, some students may not know this. Also, are we assuming students know what is the RMSD? The first lecture starts from very simple concepts (what is a protein), so I would add a concise definition of RSMD on this slide or **slide 13**. This would ensure all students can follow along, even without a lecture.
-  (Optional) **Slide 17** includes "things to worry about", I think it would be good to add a sentence about what is the impact of not considering these things properly in the docking calculations.
- On **slide 20** I would highlight the tools that are open source
- (Optional) On slide 21 maybe list 2 advantages of ML-based docking over traditional methods (and 2 disadvantages)
- Overall I really like the graphical explanations of this Lecture :heart:
