### 2024-11-24 CCPBioSim/BioSim-analysis-workshop#8 @raquellrios
Comments L4
Comments:

- On **slide 8** in  "Solvate in water and add ions", I would add that the system is (normally) neutralised.
- On **slide 11** I would add a short sentence about why PBC are needed (e.g. reduce finite size effects). Additionally, on this slide I would also add a Why? question to semi-isotropic pressure coupling in membranes so that students are encouraged to think about why this might be.
- (Optional) On **slide 12** add that these timesteps values are for all atomistic simulations.
