### 2024-11-24 CCPBioSim/BioSim-analysis-workshop#4 @raquellrios
I see now that the target audience is included in the paper.md Could this be added of the description in the README file of the repo? I think it would be very useful!
