### 2024-11-24 CCPBioSim/BioSim-analysis-workshop#8 @raquellrios
Comments on P4:

- The first step of this practical is the topology. However the concept of topology is not really introduced in L4. Therefore, I believe an explanation of topology should be included in the tutorial (or lecture). It is a very key concept of MD simulations so I think it should be briefly explained in one of the two.
- As mentioned by the other reviewer, only `lig27_smiles = "[H:7][C@:6]1([C:13](=[C:11]([C:9](=[O:10])[O:8]1)[O:12][H:19])[O:14][H:20])[C@:3]([H:4])([C:2]([H:16])([H:17])[O:1][H:15])[O:5][H:18]"` works
- When TIP3P is mentioned after cell 11, I would just add that this is a three point charge water model.
