### 2024-11-24 CCPBioSim/BioSim-analysis-workshop#7 @raquellrios
Practical 3:

- (Optional) I think adding a visualization cell of the protein and ligand that the students are going to dock after getting it from the PDB would enhance the students' understanding of the steps they are taking.
- Really like the "Questions to think about"  and visualization of the docking grid :heart:
