### 2024-11-24 #265 @raquellrios
I am linking here my comments for the first part:

- Repo intro: https://github.com/CCPBioSim/BioSim-analysis-workshop/issues/4
- L1: https://github.com/CCPBioSim/BioSim-analysis-workshop/issues/5
- L2: https://github.com/CCPBioSim/BioSim-analysis-workshop/issues/6
- L3 and P3: https://github.com/CCPBioSim/BioSim-analysis-workshop/issues/7
- L4 and P4: https://github.com/CCPBioSim/BioSim-analysis-workshop/issues/8
