### 2024-11-25 CCPBioSim/BioSim-analysis-workshop#9 @raquellrios
Comments L5
Overall, this lecture is really clear with loads of visual representations. I only have one optional comment regarding two slides, which I have added thinking of students reading these slides without the lecture.

- (Optional) The pairwise RMSD slide (**slide 8**) has a bullet point explaining why this metric is useful. However  **slide 9** (RMSF) and ** slide 10 ** (rgyr) do not. I think explicitly stating an example of why these metrics are useful (e.g. RMSD --> find flexible regions in a protein) would help students follow the lecture slides.
