### 2024-11-25 CCPBioSim/BioSim-analysis-workshop#9 @raquellrios
Comment on P5:

I really like this practical session overall. It contains detailed explanations and a lot of "print statements" which will help students understand what they are doing. The ngl viewer did not work properly for me in this tutorial.
