### 2024-11-25 CCPBioSim/BioSim-analysis-workshop#10 @raquellrios
Comments L6
Comments:

- (Optional) I think to ease the transitions between **slide 15** and **slide 16**, I would include a bullet point with the limitations of PCA in MD (e.g. capturing certain slow dynamics). This can be used as a motivation to use tICA.
- (Optional) on **slide 16** tICA is really well introduced, however adding a bullet point with examples of transitions that tICA can identify would allow students to better understand its relevance.
- Overall I think this lecture manages to explain very complex concepts in a very easy-to-follow form with the aid of great visuals :heart:
