### 2024-11-25 CCPBioSim/BioSim-analysis-workshop#10 @raquellrios
I really like both tutorials on dimensionality reduction. I think they explain really clearly the concept of dimesionality reduction with PCA and tICA, the difference between them and their application to MD simulations. Also all visualizations worked for me in these tutorials.
