### 2024-11-25 CCPBioSim/BioSim-analysis-workshop#11 @raquellrios
Comments L7
This lecture is concise and clear. I only have one optional comment:

- (Optional) On **slide 10** adding a bullet point on when to use Density-based vs spectral clustering (e.g. dataset size, noise ) could help students (particularly those following on their own the lecture slides) better understand the difference between both methods.
