### 2024-11-25 CCPBioSim/BioSim-analysis-workshop#11 @raquellrios
Comments on P7:

- I really like this tutorial, very clear visual representations of the difference between DBSCAN and k-clustering and how they can be applied to simulation data.
