### 2024-11-25 CCPBioSim/BioSim-analysis-workshop#12 @raquellrios
Comments L8
This lecture provides a good introduction to supervised classification, introducing various algorithms. I only have one comment:

- Adding a bullet point stating recommendations on when the use NN, RF or DT (e.g. dataset size, robustness to noise) would improve the students' understanding of the methods.
