### 2024-11-25 CCPBioSim/BioSim-analysis-workshop#12 @raquellrios
Comments on P8:

- I do not have any comments for this practical, it looks great. It clearly shows how to use extract features of a MD simulations and how to use a random forest to classify the MD conformations of the trajectory.
