### 2024-11-25 #265 @raquellrios
I am linking here my comments for the second part:

- L5 and P5: https://github.com/CCPBioSim/BioSim-analysis-workshop/issues/9
- L6 and P6: https://github.com/CCPBioSim/BioSim-analysis-workshop/issues/10
- L7 and P7: https://github.com/CCPBioSim/BioSim-analysis-workshop/issues/11
- L8 and P8: https://github.com/CCPBioSim/BioSim-analysis-workshop/issues/12
