### 2024-11-25 #265 @raquellrios
**Review:**

The course "Molecular Dynamics Simulation and Analysis Workshop", developed by @ppxasjsm and @degiacom covers the theoretical foundations of MD simulations of protein systems and their analysis, including advanced analysis topics, such as dimensionality reduction and supervised machine learning (ML) classifications. The course targets graduate students with a basic understanding of MD simulations and Python.

The course is divided into two units:

- Unit 1 (Lectures 1–4): focuses on protein simulation preparation, including essential steps like protein structure preparation, or protein-ligand docking.
- Unit 2 (Lectures 5–8): centers around simulation analysis, incorporating modern techniques like dimensionality reduction and ML-based classification.

Each lecture includes a PDF document with slides. I found the content in all lectures clear and easy to follow, enhanced by numerous visual aids that will help students understand complex topics. Such visual representations will be of great help for students following this material individually. Furthermore, the analysis lectures (Unit 2) provide examples that demonstrate the practical application of the described methods (PCA, Random forests, etc.) to molecular simulations.

In addition to the slides, the course offers practical sessions (except for Lecture 1, which serves as an introduction to proteins). Most of these sessions are presented as Jupyter Notebooks that run seamlessly on Google Colab. These notebooks are well-structured with clear code explanations. Furthermore, practicals include questions to students. These questions will foster active learning, preventing students from just pressing "run" on every cell without thinking about what they are doing. The inclusion of solutions to these questions ensures that students can complete the exercises independently. The use of Colab eliminates the need for students to download software or have access to high-performance computing resources, significantly improving the usability and accessibility of the course. Also, a key strength of the course is its exclusive reliance on open-source software, aligning with the principles of open science and ensuring accessibility for all students.

Furthermore, the course is able to go from an introduction to proteins to ML, addressing both fundamental and cutting-edge topics like AlphaFold and supervised classification. I believe Lectures 6–8 are particularly noteworthy, as they cover advanced topics for which there are not that many well established learning resources (compared to docking or MD simulations) that cover these topics in such a clear way. Also, the modular design of the material will allow teachers to adopt individual lectures or practical sessions independently, except for Practical 8, which builds upon Practical 7. This flexibility increases the course's applicability across different educational settings.

Overall I only have a few comments on the material, which I have reported as Issues in the target repository. Some of these comments are classified as "optional", which means that I do not think they are absolutely necessary but would improve explainability. Regarding the JOSE checklist, I do not see "community guidelines" in the documentation, but I assume this is a self-contained module so it may not need them.

This course provides a robust, well-designed learning process that spans foundational to advanced topics in MD simulations. Its focus on modern methods makes it a valuable resource for students and teachers. I highly recommend this material for publication after addressing the submitted comments.
