### 2024-12-17 CCPBioSim/BioSim-analysis-workshop#4 @degiacom
Thank you for your feedback! Adding target audience information in README.md file is definitely useful, this has now been done. Concerning the typo, we were going for the plural of "analysis".
