### 2024-12-18 be0efd7
Update Lecture_1_Introduction.pdf
Addressing comments in Issue #5
