### 2024-12-18 CCPBioSim/BioSim-analysis-workshop#5 @degiacom
Thank you for the suggestions, all have been implemented in [be0efd7](https://github.com/CCPBioSim/BioSim-analysis-workshop/commit/be0efd777cf6be0c514a7cff224d8dc4a8762c0c). Concerning intrinsically disordered proteins, we added few slides at the end of the lecture, also including the idea that proteins occupy a conformational space. We feel this constitutes a nice hook to the following lectures.
