### 2024-12-18 055810f
Updated Lecture 2
addressed comments in Issue #6
