+++ 2_Protein_Preparation/2_Protein_Preparation.txt
@@ -1,1 +1,0 @@
-TODO: Create worksheet for tasks with protein prep
