### 2024-12-18 CCPBioSim/BioSim-analysis-workshop#6 @degiacom
Thank you for the comments, the slides have been updated in [055810f](https://github.com/CCPBioSim/BioSim-analysis-workshop/commit/055810f2f7f43572503fad26f5607d98e2d546b0).
