### 2024-12-18 4f5f70b
Update Lecture_4_Simulation_setup.pdf
Addressed issue #8
