### 2024-12-18 CCPBioSim/BioSim-analysis-workshop#8 @degiacom
Thank you for the feedback. Comments on slides are addressed in [4f5f70b](https://github.com/CCPBioSim/BioSim-analysis-workshop/commit/4f5f70bb3d39bc0a0e2835fba57b8c2aaba290a9)
