### 2024-12-18 946d803
Update Lecture_7_Clustering.pdf
Addressing comments in #11
