### 2024-12-18 CCPBioSim/BioSim-analysis-workshop#11 @degiacom
Thank you for your feedback, slides have been updated in [946d803](https://github.com/CCPBioSim/BioSim-analysis-workshop/commit/946d803e1b3eee6a6f66d75f8366aae0ffa6c2ea)
