### 2024-12-18 CCPBioSim/BioSim-analysis-workshop#9 @degiacom
Thank you for the suggestion, the slides have been updated in [c0aaf73](https://github.com/CCPBioSim/BioSim-analysis-workshop/commit/c0aaf732f0eaf0dcbdbf4f30874a23a14ba176a0)
