### 2024-12-20 CCPBioSim/BioSim-analysis-workshop#6 @ppxasjsm
And in issue [93f81a0](https://github.com/CCPBioSim/BioSim-analysis-workshop/commit/93f81a03d2ceee9b18b6668e3a8f6b555b62fd88)
