### 2024-12-20 CCPBioSim/BioSim-analysis-workshop#7 @ppxasjsm
The following changes have been done:
- Added two slides explaining RMSD/pose quality and score quality (around Slide 15)
- Added an extra point to slide 7 commenting on ligand similarity.
- I agree the image on slide 9 is not great but it is taken from a paper and I don't want to redo it. I have added a citation on the slide now though.
- Slide 17 I don't know how to add this in a sensible way on the slide but would discuss it in a lecture.
- Added highlight to open-source tools
- I think there is no real consensus abot what you ask for slide 21 so I will leave it as is.
