### 2024-12-20 3485bf6
Updated L3
addressing issue  #7
