### 2024-12-20 CCPBioSim/BioSim-analysis-workshop#7 @ppxasjsm
This is the commit that fixes these [3485bf6](https://github.com/CCPBioSim/BioSim-analysis-workshop/commit/3485bf658fdc5b2e699330d1dd32af0dee6cdd01)
