### 2024-12-21 CCPBioSim/BioSim-analysis-workshop#9 @ppxasjsm
I think the NGL viewer is an issue as it will break with updates all the time. I will leave it as is for now as it is a game of catch up.
