### 2024-12-21 CCPBioSim/BioSim-analysis-workshop#10 @ppxasjsm
I like your optional suggestions I think for the sake of the paper release I would leave it and we may make some improvements when we rerun the course in the summer. I'll leave the issue open as a reminder but would assume this has been addressed for the review.
