### 2024-12-22 a90c35b
Update Lecture_8_classification.pdf
Addressing issue #12
