### 2024-12-22 CCPBioSim/BioSim-analysis-workshop#12 @degiacom
Thank you for the idea! A comparison between NN and RF has been added in [a90c35b](https://github.com/CCPBioSim/BioSim-analysis-workshop/commit/a90c35b7a77157442b5996c13b60b1c518c02697)
