### 2024-12-22 abd7e9d
Update Lecture_6_DR.pdf
addressing issue #10
