### 2024-12-22 CCPBioSim/BioSim-analysis-workshop#10 @degiacom
I have added one slide explaining the difference between large-scale motions and slow kinetics in [abd7e9d](https://github.com/CCPBioSim/BioSim-analysis-workshop/commit/abd7e9d942f8f3d39e72d7d96b419b32f95b5c0a)
