### 2024-12-22 CCPBioSim/BioSim-analysis-workshop#8 @ppxasjsm
I have added explanation on TIP3P as well as linked multiple methods for going from smiles string to molecule. The issue with reading the smiles from file via .smi I cannot recreate either in colab nor my local install. Everything is in this commit https://github.com/CCPBioSim/BioSim-analysis-workshop/commit/cd8c7f7b8053c89d6744f7fa06beddf9e2490828
