### 2024-12-22 #265 @ppxasjsm
Hi @djcole56,

>  In 4_Simulation_Setup, only this method works:
>
> lig27_smiles = "[H:7][C@:6]1([C:13](=[C:11]([C:9](=[O:10])[O:8]1)[O:12][H:19])[O:14][H:20])[C@:3]([H:4])([C:2]([H:16])([H:17])[O:1][H:15])[O:5][H:18]"
> ligand = Molecule.from_mapped_smiles(lig27_smiles)
> ie reading from file does not (I guess because it's not a mapped smiles).
>
> More generally, it would be good to include instruction for how to make a mapped smiles if this is necessary input.

I cannot recreate your issues. Maybe you could try again? I have also added a link to more explantions in an other resource on [how to read in molcules](https://docs.openforcefield.org/projects/toolkit/en/stable/users/molecule_cookbook.html). Please take a look a this commit to see if everything has been adequately addressed: https://github.com/CCPBioSim/BioSim-analysis-workshop/commit/cd8c7f7b8053c89d6744f7fa06beddf9e2490828
