### 2024-12-22 f153e9e
Update Lecture_4_Simulation_setup.pdf
addressing issue #8
