### 2024-12-22 CCPBioSim/BioSim-analysis-workshop#8 @ppxasjsm
We added a slide on topology to lecture 4. see [f153e9e](https://github.com/CCPBioSim/BioSim-analysis-workshop/commit/f153e9e9f8b84daa397721ff31efeeec30f26170). I believe this addresses all the comments.
