### 2024-12-22 CCPBioSim/BioSim-analysis-workshop#12 @ppxasjsm
I belive we have now addressed everything
