### 2024-12-22 #265 @ppxasjsm
Hi @raquellrios and @djcole56 thank you so much for your nice comments and thoughtful reviews. We have not added a statement of need to the Github as it is covered in the paper. We have added some community guidelines in this commit: https://github.com/CCPBioSim/BioSim-analysis-workshop/commit/d381349b4aafaff0cf0713c417d4bd00c36257b3. I believe we have now addressed all the concerns raised by issues on the repo and in this thread. If you agree, then we hope that @arm61 can proceed with the submission. Have a lovely holiday break, everyone!
