### 2024-12-22 d381349
adding community guidelines
