### 2024-12-28 #265 @djcole56
- I can no longer reproduce the issue in workshop 4, and further help has been added in any case.
- Community guidelines have been added.

@arm61 all good to go from my side.
