### 2025-01-20 #265 @arm61
Sorry for the delay all (in short -- teaching). @ppxasjsm and @degiacom, is there a nicer way to format the "Open in Colab" parts in unit breakdowns, perhaps more in keeping with the style of the pdf? Just a thought.
