### 2025-01-20 #265 @degiacom
Hello @arm61, thank you for having had a look at our paper. Perhaps, we could leave the Colab badges in the repo's README file, but replace them with a text ("Open in Colab") with hyperlink in the paper?
