### 2025-01-21 98c4d60
Substituted Colab badge with text for better formatting
