### 2025-01-21 702bd49
Fixed links to material in schedule
