+++ README.md
@@ -13,13 +13,13 @@ The course is constituted of lectures (L1-8) and practical (P) sessions, subdivi
 
 | Session                            | Materials |
 |------------------------------------|-----------|
-| L1: Introduction to Proteins | [Lecture Slides](https://github.com/CCPBioSim/Into_to_MD_simulation_and_analysis/blob/main/1_Introduction/Lecture_1_Introduction.pdf) | 
-| L2: Understanding Protein Systems | [Lecture Slides](https://github.com/CCPBioSim/Into_to_MD_simulation_and_analysis/blob/main/2_Protein_Preparation/Lecture_2_Protein_Prep.pdf)
+| L1: Introduction to Proteins | [Lecture Slides](1_Introduction/Lecture_1_Introduction.pdf) | 
+| L2: Understanding Protein Systems | [Lecture Slides](2_Protein_Preparation/Lecture_2_Protein_Prep.pdf)
 | P: Understanding Protein Systems, contd. | [Webserver](https://server.poissonboltzmann.org/pdb2pqr)|
-| L3: Protein-Ligand Docking                 |  [Lecture Slides](https://github.com/CCPBioSim/Into_to_MD_simulation_and_analysis/blob/main/3_Docking/Lecture_3_Docking.pdf)| 
-| P: Protein-Ligand Docking                 |  [![Docking](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/CCPBioSim/Into_to_MD_simulation_and_analysis/blob/main/3_Docking/3_Docking.ipynb)| 
-| L4: Simulation Setup          | [Lecture Slides](https://github.com/CCPBioSim/Into_to_MD_simulation_and_analysis/blob/main/4_Simulation_Setup/Lecture_4_Simulation_setup.pdf) |
-| P: Simulation Setup          | [![Simulation](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/CCPBioSim/Into_to_MD_simulation_and_analysis/blob/main/4_Simulation_Setup/4_Simulation_Setup.ipynb) |
+| L3: Protein-Ligand Docking                 |  [Lecture Slides](3_Docking/Lecture_3_Docking.pdf)| 
+| P: Protein-Ligand Docking                 |  [![Docking](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/CCPBioSim/BioSim-analysis-workshop/blob/main/3_Docking/3_Docking.ipynb)| 
+| L4: Simulation Setup          | [Lecture Slides](4_Simulation_Setup/Lecture_4_Simulation_setup.pdf) |
+| P: Simulation Setup          | [![Simulation](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/CCPBioSim/BioSim-analysis-workshop/blob/main/4_Simulation_Setup/4_Simulation_Setup.ipynb) |
 
 
 ### Unit 2: Simulation Analysis
@@ -28,14 +28,14 @@ The course is constituted of lectures (L1-8) and practical (P) sessions, subdivi
 | Session                                             | Materials |
 |-----------------------------------------------------|-----------|
 | L5: Simulation Basic Analyses             | [Lecture Slides](5_Analysis_MDAnalysis/Lecture_5_Analysis_MDAnalysis.pdf)|
[…]
