### 2025-01-21 a4a9f5d
Fixed hyperlinks in schedule
