### 2025-01-21 #265 @degiacom
Done! I replaced the "Open in Colab" badges with a "Notebook" plain text.
