### 2025-01-22 439993f
Update paper.md
minor grammatical and orthographic fixes
