### 2025-01-22 #265 @degiacom
We have now checked grammar/orthograph of the paper, we are happy with it!
