### 2025-01-26 #265 @arm61
Fab, can you now mint a DOI using Zenodo for the manuscript and make a versioned release of the materials?
