### 2025-01-29 #265 @ppxasjsm
Hi @arm61, the DOI from Zenodo is: 10.5281/zenodo.14767012, the version is: 1.0.0 Anything else needed now?
