### 2025-01-29 #265 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
