### 2025-05-01 #265 @labarba
hi @degiacom 👋  – The Zenodo archive has the title "BioSim-analysis-workshop" while the paper is titled "A course on the setup, running, and analysis of biomolecular simulations" … and the repo has a third title! "Molecular Dynamics Simulation and Analysis workshop"

We request that authors edit the metadata of the Zenodo deposit so title matches the JOSE paper. It's just cleaner that way as readers see these as part of the "same scholarly object." Could you do that? Pick one title and apply it consistently for this release and archive.
