### 2025-05-04 6c4031c
small fixes paper.md
