+++ paper/paper.md
@@ -52,7 +52,7 @@ This is a graduate-level course, aimed at beginners in biomolecular simulation. 
 
 ## Content
 
-The objective of this course is not to make students proficient in one or a few specific software tools e for MD simulation preparation, execution, or analysis. Instead, it is aimed at providing students with a general overview of the key decision-making required to carry out MD simulations of biomolecules and extracting quantitative data from them. In this context, the course is divided into two Units featuring practical sessions and lectures. Practical sessions demonstrate how key concepts in molecular modelling are put into practice by exposing student to authentic tasks leveraging on commonly used Python packages, such as MDAnalysis [@michaud-agrawal2011mdanalysis] [@oliver_beckstein-proc-scipy-2016] [@alibay2023building] and scikit-learn [@pedregosa2011scikitlearn]. Lectures are software-agnostic and provide additional material to the course. While each practical session can be run by students on their own computer, these are also available in Google colab. This solution, requiring no local installation, is especially suitable for those unfamiliar with setting up a Python environment, or having limited access to computational resources.
+The objective of this course is not to make students proficient in one or a few specific software tools for MD simulation preparation, execution, or analysis. Instead, it is aimed at providing students with a general overview of the key decision-making required to carry out MD simulations of biomolecules and extracting quantitative data from them. In this context, the course is divided into two Units featuring practical sessions and lectures. Practical sessions demonstrate how key concepts in molecular modelling are put into practice by exposing students to authentic tasks leveraging on commonly used Python packages, such as MDAnalysis [@michaud-agrawal2011mdanalysis,@oliver_beckstein-proc-scipy-2016,@alibay2023building] and scikit-learn [@pedregosa2011scikitlearn]. Lectures are software-agnostic and provide additional material to the course. While each practical session can be run by students on their own computer, these are also available in Google Colab. This solution, requiring no local installation, is especially suitable for those unfamiliar with setting up a Python environment, or having limited access to computational resources.
 
 ### Unit 1: Simulation Preparation
 
