### 2025-05-04 #265 @labarba
I made some small edits here: https://github.com/CCPBioSim/BioSim-analysis-workshop/pull/13
